randomly fix everything, add id creation
