+++ config/data_processing/fixes.csv
@@ -1,2 +1,2 @@
-﻿id,transaction_id,date,description,info,amount,class,category,sub_category
-f1,id,pvm,kauppa,,summa,pois,syy miksi korjattu,syy miksi korjattu
+﻿id,date,description,info,amount,class,category,sub_category,transaction_id
+f1,2023-01-30,muutettu pvm ja description esimerkiksi (täytetyt kentät muuttuvat),,,,,,20230131__varat__asunto____1000000__actual__varatcsv__mkk
