+++ data/intermediate/8_targets.csv
@@ -1,0 +1,49 @@
+date,account,description,info,amount_original,class,category,sub_category,rule_id,source_file,row_type,share,amount,owner,split,not_unique_id
+2023-01-31,Target,Säästötavoite 1,,1000,tulot,,,Säästötavoite 1,targets.csv,Target,1,1000,mkk,False,20230131__target__säästötavoite1__1000__target__mkk
+2023-02-28,Target,Säästötavoite 1,,1000,tulot,,,Säästötavoite 1,targets.csv,Target,1,1000,mkk,False,20230228__target__säästötavoite1__1000__target__mkk
+2023-03-31,Target,Säästötavoite 1,,1000,tulot,,,Säästötavoite 1,targets.csv,Target,1,1000,mkk,False,20230331__target__säästötavoite1__1000__target__mkk
+2023-04-30,Target,Säästötavoite 1,,1000,tulot,,,Säästötavoite 1,targets.csv,Target,1,1000,mkk,False,20230430__target__säästötavoite1__1000__target__mkk
+2023-05-31,Target,Säästötavoite 1,,1000,tulot,,,Säästötavoite 1,targets.csv,Target,1,1000,mkk,False,20230531__target__säästötavoite1__1000__target__mkk
+2023-06-30,Target,Säästötavoite 1,,1000,tulot,,,Säästötavoite 1,targets.csv,Target,1,1000,mkk,False,20230630__target__säästötavoite1__1000__target__mkk
+2023-07-31,Target,Säästötavoite 1,,1000,tulot,,,Säästötavoite 1,targets.csv,Target,1,1000,mkk,False,20230731__target__säästötavoite1__1000__target__mkk
+2023-08-31,Target,Säästötavoite 1,,1000,tulot,,,Säästötavoite 1,targets.csv,Target,1,1000,mkk,False,20230831__target__säästötavoite1__1000__target__mkk
+2023-09-30,Target,Säästötavoite 1,,1000,tulot,,,Säästötavoite 1,targets.csv,Target,1,1000,mkk,False,20230930__target__säästötavoite1__1000__target__mkk
+2023-10-31,Target,Säästötavoite 1,,1000,tulot,,,Säästötavoite 1,targets.csv,Target,1,1000,mkk,False,20231031__target__säästötavoite1__1000__target__mkk
+2023-11-30,Target,Säästötavoite 1,,1000,tulot,,,Säästötavoite 1,targets.csv,Target,1,1000,mkk,False,20231130__target__säästötavoite1__1000__target__mkk
+2023-12-31,Target,Säästötavoite 1,,1000,tulot,,,Säästötavoite 1,targets.csv,Target,1,1000,mkk,False,20231231__target__säästötavoite1__1000__target__mkk
+2024-01-31,Target,Säästötavoite 1,,1000,tulot,,,Säästötavoite 1,targets.csv,Target,1,1000,mkk,False,20240131__target__säästötavoite1__1000__target__mkk
+2024-02-29,Target,Säästötavoite 1,,1000,tulot,,,Säästötavoite 1,targets.csv,Target,1,1000,mkk,False,20240229__target__säästötavoite1__1000__target__mkk
+2024-03-31,Target,Säästötavoite 1,,1000,tulot,,,Säästötavoite 1,targets.csv,Target,1,1000,mkk,False,20240331__target__säästötavoite1__1000__target__mkk
+2024-04-30,Target,Säästötavoite 1,,1000,tulot,,,Säästötavoite 1,targets.csv,Target,1,1000,mkk,False,20240430__target__säästötavoite1__1000__target__mkk
+2024-05-31,Target,Säästötavoite 1,,1000,tulot,,,Säästötavoite 1,targets.csv,Target,1,1000,mkk,False,20240531__target__säästötavoite1__1000__target__mkk
+2024-06-30,Target,Säästötavoite 1,,1000,tulot,,,Säästötavoite 1,targets.csv,Target,1,1000,mkk,False,20240630__target__säästötavoite1__1000__target__mkk
+2024-07-31,Target,Säästötavoite 1,,1000,tulot,,,Säästötavoite 1,targets.csv,Target,1,1000,mkk,False,20240731__target__säästötavoite1__1000__target__mkk
+2024-08-31,Target,Säästötavoite 1,,1000,tulot,,,Säästötavoite 1,targets.csv,Target,1,1000,mkk,False,20240831__target__säästötavoite1__1000__target__mkk
+2024-09-30,Target,Säästötavoite 1,,1000,tulot,,,Säästötavoite 1,targets.csv,Target,1,1000,mkk,False,20240930__target__säästötavoite1__1000__target__mkk
+2024-10-31,Target,Säästötavoite 1,,1000,tulot,,,Säästötavoite 1,targets.csv,Target,1,1000,mkk,False,20241031__target__säästötavoite1__1000__target__mkk
+2024-11-30,Target,Säästötavoite 1,,1000,tulot,,,Säästötavoite 1,targets.csv,Target,1,1000,mkk,False,20241130__target__säästötavoite1__1000__target__mkk
+2024-12-31,Target,Säästötavoite 1,,1000,tulot,,,Säästötavoite 1,targets.csv,Target,1,1000,mkk,False,20241231__target__säästötavoite1__1000__target__mkk
+2023-01-31,Target,Säästötavoite 1,,-500,menot,,,Säästötavoite 1,targets.csv,Target,1,-500,mkk,False,20230131__target__säästötavoite1__500__target__mkk
+2023-02-28,Target,Säästötavoite 1,,-500,menot,,,Säästötavoite 1,targets.csv,Target,1,-500,mkk,False,20230228__target__säästötavoite1__500__target__mkk
+2023-03-31,Target,Säästötavoite 1,,-500,menot,,,Säästötavoite 1,targets.csv,Target,1,-500,mkk,False,20230331__target__säästötavoite1__500__target__mkk
+2023-04-30,Target,Säästötavoite 1,,-500,menot,,,Säästötavoite 1,targets.csv,Target,1,-500,mkk,False,20230430__target__säästötavoite1__500__target__mkk
+2023-05-31,Target,Säästötavoite 1,,-500,menot,,,Säästötavoite 1,targets.csv,Target,1,-500,mkk,False,20230531__target__säästötavoite1__500__target__mkk
+2023-06-30,Target,Säästötavoite 1,,-500,menot,,,Säästötavoite 1,targets.csv,Target,1,-500,mkk,False,20230630__target__säästötavoite1__500__target__mkk
+2023-07-31,Target,Säästötavoite 1,,-500,menot,,,Säästötavoite 1,targets.csv,Target,1,-500,mkk,False,20230731__target__säästötavoite1__500__target__mkk
+2023-08-31,Target,Säästötavoite 1,,-500,menot,,,Säästötavoite 1,targets.csv,Target,1,-500,mkk,False,20230831__target__säästötavoite1__500__target__mkk
+2023-09-30,Target,Säästötavoite 1,,-500,menot,,,Säästötavoite 1,targets.csv,Target,1,-500,mkk,False,20230930__target__säästötavoite1__500__target__mkk
+2023-10-31,Target,Säästötavoite 1,,-500,menot,,,Säästötavoite 1,targets.csv,Target,1,-500,mkk,False,20231031__target__säästötavoite1__500__target__mkk
+2023-11-30,Target,Säästötavoite 1,,-500,menot,,,Säästötavoite 1,targets.csv,Target,1,-500,mkk,False,20231130__target__säästötavoite1__500__target__mkk
+2023-12-31,Target,Säästötavoite 1,,-500,menot,,,Säästötavoite 1,targets.csv,Target,1,-500,mkk,False,20231231__target__säästötavoite1__500__target__mkk
+2024-01-31,Target,Säästötavoite 1,,-500,menot,,,Säästötavoite 1,targets.csv,Target,1,-500,mkk,False,20240131__target__säästötavoite1__500__target__mkk
+2024-02-29,Target,Säästötavoite 1,,-500,menot,,,Säästötavoite 1,targets.csv,Target,1,-500,mkk,False,20240229__target__säästötavoite1__500__target__mkk
+2024-03-31,Target,Säästötavoite 1,,-500,menot,,,Säästötavoite 1,targets.csv,Target,1,-500,mkk,False,20240331__target__säästötavoite1__500__target__mkk
+2024-04-30,Target,Säästötavoite 1,,-500,menot,,,Säästötavoite 1,targets.csv,Target,1,-500,mkk,False,20240430__target__säästötavoite1__500__target__mkk
+2024-05-31,Target,Säästötavoite 1,,-500,menot,,,Säästötavoite 1,targets.csv,Target,1,-500,mkk,False,20240531__target__säästötavoite1__500__target__mkk
+2024-06-30,Target,Säästötavoite 1,,-500,menot,,,Säästötavoite 1,targets.csv,Target,1,-500,mkk,False,20240630__target__säästötavoite1__500__target__mkk
+2024-07-31,Target,Säästötavoite 1,,-500,menot,,,Säästötavoite 1,targets.csv,Target,1,-500,mkk,False,20240731__target__säästötavoite1__500__target__mkk
+2024-08-31,Target,Säästötavoite 1,,-500,menot,,,Säästötavoite 1,targets.csv,Target,1,-500,mkk,False,20240831__target__säästötavoite1__500__target__mkk
+2024-09-30,Target,Säästötavoite 1,,-500,menot,,,Säästötavoite 1,targets.csv,Target,1,-500,mkk,False,20240930__target__säästötavoite1__500__target__mkk
+2024-10-31,Target,Säästötavoite 1,,-500,menot,,,Säästötavoite 1,targets.csv,Target,1,-500,mkk,False,20241031__target__säästötavoite1__500__target__mkk
+2024-11-30,Target,Säästötavoite 1,,-500,menot,,,Säästötavoite 1,targets.csv,Target,1,-500,mkk,False,20241130__target__säästötavoite1__500__target__mkk
+2024-12-31,Target,Säästötavoite 1,,-500,menot,,,Säästötavoite 1,targets.csv,Target,1,-500,mkk,False,20241231__target__säästötavoite1__500__target__mkk
