+++ data/intermediate/debug/3_append_2_after_types.csv
@@ -1,0 +1,25 @@
+date,account,description,info,amount,row_type,source_file
+2023-01-31,velat,asuntolaina,,-90000.0,Actual,velat.csv
+2023-02-28,velat,asuntolaina,,-89500.0,Actual,velat.csv
+2023-03-31,velat,asuntolaina,,-89000.0,Actual,velat.csv
+2023-04-30,velat,asuntolaina,,-88500.0,Actual,velat.csv
+2023-05-31,velat,asuntolaina,,-88000.0,Actual,velat.csv
+2023-06-30,velat,asuntolaina,,-87500.0,Actual,velat.csv
+2023-07-31,velat,asuntolaina,,-87000.0,Actual,velat.csv
+2023-08-31,velat,asuntolaina,,-86500.0,Actual,velat.csv
+2023-09-30,velat,asuntolaina,,-86000.0,Actual,velat.csv
+2023-10-31,velat,asuntolaina,,-85500.0,Actual,velat.csv
+2023-11-30,velat,asuntolaina,,-85000.0,Actual,velat.csv
+2023-12-31,velat,asuntolaina,,-84500.0,Actual,velat.csv
+2024-01-31,velat,asuntolaina,,-84000.0,Actual,velat.csv
+2024-02-28,velat,asuntolaina,,-83500.0,Actual,velat.csv
+2024-03-31,velat,asuntolaina,,-83000.0,Actual,velat.csv
+2024-04-30,velat,asuntolaina,,-82500.0,Actual,velat.csv
+2024-05-31,velat,asuntolaina,,-82000.0,Actual,velat.csv
+2024-06-30,velat,asuntolaina,,-81500.0,Actual,velat.csv
+2024-07-31,velat,asuntolaina,,-81000.0,Actual,velat.csv
+2024-08-31,velat,asuntolaina,,-80500.0,Actual,velat.csv
+2024-09-30,velat,asuntolaina,,-80000.0,Actual,velat.csv
+2024-10-31,velat,asuntolaina,,-79500.0,Actual,velat.csv
+2024-11-30,velat,asuntolaina,,-79000.0,Actual,velat.csv
+2024-12-31,velat,asuntolaina,,-78500.0,Actual,velat.csv
