--- data/intermediate/debug/append_2_after_types.csv
@@ -1,25 +1,0 @@
-date,account,description,info,amount,record_type,source_file,transaction_id
-2023-01-31,velat,asuntolaina,,-90000.0,Actual,velat.csv,20230131__velat__asuntolaina____900000__actual__velatcsv
-2023-02-28,velat,asuntolaina,,-89500.0,Actual,velat.csv,20230228__velat__asuntolaina____895000__actual__velatcsv
-2023-03-31,velat,asuntolaina,,-89000.0,Actual,velat.csv,20230331__velat__asuntolaina____890000__actual__velatcsv
-2023-04-30,velat,asuntolaina,,-88500.0,Actual,velat.csv,20230430__velat__asuntolaina____885000__actual__velatcsv
-2023-05-31,velat,asuntolaina,,-88000.0,Actual,velat.csv,20230531__velat__asuntolaina____880000__actual__velatcsv
-2023-06-30,velat,asuntolaina,,-87500.0,Actual,velat.csv,20230630__velat__asuntolaina____875000__actual__velatcsv
-2023-07-31,velat,asuntolaina,,-87000.0,Actual,velat.csv,20230731__velat__asuntolaina____870000__actual__velatcsv
-2023-08-31,velat,asuntolaina,,-86500.0,Actual,velat.csv,20230831__velat__asuntolaina____865000__actual__velatcsv
-2023-09-30,velat,asuntolaina,,-86000.0,Actual,velat.csv,20230930__velat__asuntolaina____860000__actual__velatcsv
-2023-10-31,velat,asuntolaina,,-85500.0,Actual,velat.csv,20231031__velat__asuntolaina____855000__actual__velatcsv
-2023-11-30,velat,asuntolaina,,-85000.0,Actual,velat.csv,20231130__velat__asuntolaina____850000__actual__velatcsv
-2023-12-31,velat,asuntolaina,,-84500.0,Actual,velat.csv,20231231__velat__asuntolaina____845000__actual__velatcsv
-2024-01-31,velat,asuntolaina,,-84000.0,Actual,velat.csv,20240131__velat__asuntolaina____840000__actual__velatcsv
-2024-02-28,velat,asuntolaina,,-83500.0,Actual,velat.csv,20240228__velat__asuntolaina____835000__actual__velatcsv
-2024-03-31,velat,asuntolaina,,-83000.0,Actual,velat.csv,20240331__velat__asuntolaina____830000__actual__velatcsv
-2024-04-30,velat,asuntolaina,,-82500.0,Actual,velat.csv,20240430__velat__asuntolaina____825000__actual__velatcsv
-2024-05-31,velat,asuntolaina,,-82000.0,Actual,velat.csv,20240531__velat__asuntolaina____820000__actual__velatcsv
-2024-06-30,velat,asuntolaina,,-81500.0,Actual,velat.csv,20240630__velat__asuntolaina____815000__actual__velatcsv
-2024-07-31,velat,asuntolaina,,-81000.0,Actual,velat.csv,20240731__velat__asuntolaina____810000__actual__velatcsv
-2024-08-31,velat,asuntolaina,,-80500.0,Actual,velat.csv,20240831__velat__asuntolaina____805000__actual__velatcsv
-2024-09-30,velat,asuntolaina,,-80000.0,Actual,velat.csv,20240930__velat__asuntolaina____800000__actual__velatcsv
-2024-10-31,velat,asuntolaina,,-79500.0,Actual,velat.csv,20241031__velat__asuntolaina____795000__actual__velatcsv
-2024-11-30,velat,asuntolaina,,-79000.0,Actual,velat.csv,20241130__velat__asuntolaina____790000__actual__velatcsv
-2024-12-31,velat,asuntolaina,,-78500.0,Actual,velat.csv,20241231__velat__asuntolaina____785000__actual__velatcsv
